Woul/bp_django/requirements.txt
mlflow==2.3
opencv-python==4.7.0.72
Pillow==9.5.0
Django==4.2
tensorflow==2.12.0
openai==0.27.6
django-environ==0.4.5
django-cors-headers==3.8.0
psycopg2-binary==2.9.1
pydub==0.25.1
virtualenv
ultralytics
celery==5.3.1
azure-cognitiveservices-speech
celery
joblib>=1.1.0
numba==0.56.4
numpy==1.23.5
scipy==1.9.3
librosa==0.9.1
llvmlite==0.39.0
fairseq==0.12.2
faiss-cpu==1.7.3
gradio
Cython
pydub>=0.25.1
soundfile>=0.12.1
ffmpeg-python>=0.2.0
tensorboardX
Jinja2>=3.1.2
json5
Markdown
matplotlib>=3.7.0
matplotlib-inline>=0.1.3
praat-parselmouth>=0.4.2
Pillow>=9.1.1
resampy>=0.4.2
scikit-learn
starlette>=0.25.0
tensorboard
tqdm>=4.63.1
tornado>=6.1
Werkzeug>=2.2.3
uc-micro-py>=1.0.1
sympy>=1.11.1
tabulate>=0.8.10
PyYAML>=6.0
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
fsspec>=2022.11.0
absl-py>=1.2.0
audioread
uvicorn>=0.21.1
colorama>=0.4.5
pyworld>=0.3.2
httpx==0.23.0
onnxruntime-gpu
torchcrepe==0.0.20
eventlet
--extra-index-url https://download.pytorch.org/whl/cu117 torch==1.13.1+cu117
--extra-index-url https://download.pytorch.org/whl/cu117 torchvision==0.14.1+cu117
--extra-index-url https://download.pytorch.org/whl/cu117 torchaudio==0.13.1
djangorestframework